Created a new class for csv methods and moved the existing csv methods into the class
